Added refactoring for xenia.
