+++ app/xenia/handlers/query.go
@@ -4,9 +4,7 @@ package handlers
 import (
 	"encoding/json"
 	"net/http"
-	"net/url"
 
-	"github.com/coralproject/xenia/pkg/exec"
 	"github.com/coralproject/xenia/pkg/query"
 
 	"github.com/ardanlabs/kit/web/app"
@@ -65,52 +63,6 @@ func (queryHandle) Upsert(c *app.Context) error {
 	}
 
 	c.Respond(nil, http.StatusNoContent)
-	return nil
-}
-
-//==============================================================================
-
-// Execute runs the specified Set and return results.
-// 200 Success, 400 Bad Request, 404 Not Found, 500 Internal
-func (queryHandle) Execute(c *app.Context) error {
-	set, err := query.GetByName(c.SessionID, c.DB, c.Params["name"])
-	if err != nil {
-		if err == query.ErrNotFound {
-			err = app.ErrNotFound
-		}
-		return err
-	}
-
-	return execute(c, set)
-}
-
-// ExecuteCustom runs the provided Set and return results.
-// 200 Success, 400 Bad Request, 404 Not Found, 500 Internal
-func (queryHandle) ExecuteCustom(c *app.Context) error {
-	var set *query.Set
-	if err := json.NewDecoder(c.Request.Body).Decode(&set); err != nil {
-		return err
-	}
-
-	return execute(c, set)
-}
-
-// execute takes a context and Set and executes the set returning
-// any possible response.
-func execute(c *app.Context, set *query.Set) error {
-	var vars map[string]string
-	if c.Request.URL.RawQuery != "" {
-		if m, err := url.ParseQuery(c.Request.URL.RawQuery); err == nil {
-			vars = make(map[string]string)
-			for k, v := range m {
-				vars[k] = v[0]
-			}
-		}
-	}
-
-	result := exec.Set(c.SessionID, c.DB, set, vars)
-
-	c.Respond(result, http.StatusOK)
 	return nil
 }
