+++ app/xenia/tests/exec_test.go
@@ -1,0 +1,124 @@
+// Package endpoint implements users tests for the API layer.
+package tests
+
+import (
+	"bytes"
+	"encoding/json"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/coralproject/xenia/pkg/query/qfix"
+
+	"github.com/ardanlabs/kit/tests"
+)
+
+// TestExec tests the execution of a specific query.
+func TestExec(t *testing.T) {
+	tests.ResetLog()
+	defer tests.DisplayLog()
+
+	t.Log("Given the need to execute a specific query.")
+	{
+		url := "/1.0/exec/QTEST_basic?station_id=42021"
+		r := tests.NewRequest("GET", url, nil)
+		w := httptest.NewRecorder()
+
+		a.ServeHTTP(w, r)
+
+		t.Logf("\tWhen calling url : %s", url)
+		{
+			if w.Code != 200 {
+				t.Fatalf("\t%s\tShould be able to retrieve the query : %v", tests.Failed, w.Code)
+			}
+			t.Logf("\t%s\tShould be able to retrieve the query.", tests.Success)
+
+			recv := w.Body.String()
+			resp := `{"results":[{"Name":"Basic","Docs":[{"name":"C14 - Pasco County Buoy, FL"}]}],"error":false}`
+
+			if resp != recv {
+				t.Log(resp)
+				t.Log(recv)
+				t.Fatalf("\t%s\tShould get the expected result.", tests.Failed)
+			}
+			t.Logf("\t%s\tShould get the expected result.", tests.Success)
+		}
+	}
+}
+
+// TestExecCustom tests the execution of a custom query.
+func TestExecCustom(t *testing.T) {
+	tests.ResetLog()
+	defer tests.DisplayLog()
+
+	t.Log("Given the need to execute a custom query.")
+	{
+		qs, err := qfix.Get("basic.json")
+		if err != nil {
+			t.Fatalf("\t%s\tShould be able to retrieve the fixture : %v", tests.Failed, err)
+		}
+		t.Logf("\t%s\tShould be able to retrieve the fixture.", tests.Success)
+
+		qsStrData, err := json.Marshal(&qs)
+		if err != nil {
+			t.Fatalf("\t%s\tShould be able to marshal the fixture : %v", tests.Failed, err)
+		}
+		t.Logf("\t%s\tShould be able to marshal the fixture.", tests.Success)
+
+		url := "/1.0/exec"
+		r := tests.NewRequest("POST", url, bytes.NewBuffer(qsStrData))
+		w := httptest.NewRecorder()
+
+		a.ServeHTTP(w, r)
+
+		t.Logf("\tWhen calling url : %s", url)
+		{
+			if w.Code != 200 {
+				t.Fatalf("\t%s\tShould be able to retrieve the query : %v", tests.Failed, w.Code)
+			}
+			t.Logf("\t%s\tShould be able to retrieve the query.", tests.Success)
+
+			recv := w.Body.String()
+			resp := `{"results":[{"Name":"Basic","Docs":[{"name":"C14 - Pasco County Buoy, FL"}]}],"error":false}`
+
+			if resp != recv {
+				t.Log(resp)
+				t.Log(recv)
+				t.Fatalf("\t%s\tShould get the expected result.", tests.Failed)
+			}
+			t.Logf("\t%s\tShould get the expected result.", tests.Success)
+		}
+	}
+}
+
+// TestExecJSONP tests the execution of a specific query using JSONP.
+func TestExecJSONP(t *testing.T) {
+	tests.ResetLog()
+	defer tests.DisplayLog()
+
+	t.Log("Given the need to execute a specific query with JSONP output.")
+	{
+		url := "/1.0/exec/QTEST_basic?station_id=42021&callback=handle_data"
+		r := tests.NewRequest("GET", url, nil)
+		w := httptest.NewRecorder()
+
+		a.ServeHTTP(w, r)
+
+		t.Logf("\tWhen calling url : %s", url)
+		{
+			if w.Code != 200 {
+				t.Fatalf("\t%s\tShould be able to retrieve the query : %v", tests.Failed, w.Code)
+			}
+			t.Logf("\t%s\tShould be able to retrieve the query.", tests.Success)
+
+			recv := w.Body.String()
+			resp := `handle_data({"results":[{"Name":"Basic","Docs":[{"name":"C14 - Pasco County Buoy, FL"}]}],"error":false})`
+
+			if resp != recv {
+				t.Log(resp)
+				t.Log(recv)
+				t.Fatalf("\t%s\tShould get the expected result.", tests.Failed)
+			}
+			t.Logf("\t%s\tShould get the expected result.", tests.Success)
+		}
+	}
+}
